i fixed the main menu screen
see what happened was, the fade screen was on top of the other canvas so it was another sorting layer issue so i just changed the sorting layer and it worked... sorta. The fade function didn't cover the davis family so I couldn't just change the sorting layer so then I became programming god and programmed the shit out of it. HAHAHAHAHA take that